honggfuzz: use size_t and %zu for printing rlimit values, which can be bigger than int

+++ honggfuzz.c
@@ -87,12 +87,12 @@ static void setupRLimits(void) {
         return;
     }
     if (rlim.rlim_max < 1024) {
-        LOG_E("RLIMIT_NOFILE max limit < 1024 (%u). Expect troubles!", (unsigned int)rlim.rlim_max);
+        LOG_E("RLIMIT_NOFILE max limit < 1024 (%zu). Expect troubles!", (size_t)rlim.rlim_max);
         return;
     }
     rlim.rlim_cur = MIN(1024, rlim.rlim_max);  // we don't need more
     if (setrlimit(RLIMIT_NOFILE, &rlim) == -1) {
-        PLOG_E("Couldn't setrlimit(RLIMIT_NOFILE, cur=max=%u)", (unsigned int)rlim.rlim_max);
+        PLOG_E("Couldn't setrlimit(RLIMIT_NOFILE, cur=max=%zu)", (size_t)rlim.rlim_max);
     }
 }
